Updated aggregation of blog posts handle edge case
Generate site did not handle the the blog listings if the number of
files was less than the number of posts per page. Each blog would be
generated but the aggregate page would not be created correctly.

This commit Fixes that by checking if the number of posts are less than
the number of blogs per post and reruns the loop that generates the
aggregate page.