Added template for styleing
It seems like the only way to get styling on github pages is to let
jeykl do it for you. So there's a template file just so that I can use
my styling. I mean, I could add html headers and such to every line of
my page, but if I'm going for a simple minimilistic page. Then this will
help it along by using the template.