Reworked everything besides blog to markdown
Worked everything besides the blog into markdown. Debating about jekyll
for the blog posts. It seems like an easy way to not have to manage it
myself. Not sure yet. This is primarily a test to github to see if the
style is going to link correctly to the markdown pages.