Added resume, projects page paragraph, styles.
Added black background so the bottom of the image transitions into it
nicely. I should resize the image so that it is a good size for the
homepage and not stretched down (you should seen mountains even if you
have a normal screen size)

Added a couple paragraphs of text to the projects page to explain that I
love projects.

Ported resume page over from old site, updated some content on it, might
want to revisit later on.