Added mozzerela chicken recipe.
Bus wifi is not working right now so I can't grab the image, but once I
do I'll compile harp and push it out!

Tasty dish I made last night, looks wonderful plated and tastes just as
good. I served it with tortillini (spinach and ricotta) and it was quite
a good fit.