Created new tech-blog post.
About Authorize.ent giving me crap when trying to have a free and non
free set of lineitems and not formatting the 0 in the free one as 0.00,
even though Authorize excepts the 0 for a single item that's not a line
item.

does it make sense? no. Should it? yeah. But as all things on the
internet. We deal with it as it comes.