Added Scooter narrative.
Didn't actually mention the scooter itself in here, but the kid had one
and was balancing on it and such, but when I wrote this and reahed the
current end I thought it was a neat wrapup so I didn't continue writing
about it. Though the blue of the child's sweater matched the blue of
scooter perfectly which I thought was impressive.