Starting work on thoughts section.
Sometimes what I want to write about isn't always narratives, they're
not always poems, and they're not always smut. Sometimes I want to say
what I'm thinking, othertimes I don't. If I want to rant anonymously
then I can do so on 4chan if I want (to some extent), but if I want to
be vocal about a view and an opinion then I need an outlet on my website
to do that.

Not adding it to the main navigation yet (or the writing navigation)
until I'm done working on it.