Compiled Harp.
@GamerGateOP
Wrote a blog post on a new section of my site trying to use screenshots
of the various /v/ posts I have open right now (some of which have been
open since yesterday and the threads have died/archived).

I'm not a big site or have any following at all really, but I'm happy to
take some feedback and work and rework the post into something
constructive and useful for the gamergate community.