Fury poem.
I suppose this is my take on war. And all the fun media pushing that
happens to justify extreme cruelty and the removal of people's lives.
It's a rather negative view of the whole thing, but it's basically how I
felt after watching the tank movie fury (brad pitt).

I was stunned, I don't normally watch war movies, and the way that the
character development of Norman the "Machine" came about was done very
well. Also, reading about WHAM and other "psy-ops" tactics is
incredibly.... disturbing.