Enabled Ecotopia post and updated metadata
Also wrote in a footnote. I feel like this post is coming off a bit
angry at the concepts of women-powered government. Perhaps it is the
reading of Weston's character which puts it into the perspective but it
certainly is frustrating. Attempting to moderate my words a bit better,
saying female supremacy is perhaps a bit harsh, however how else does
one describe something in which is described as such?

The government is ran by women, the education system sees them as being
better at that and therefore training them to do so. This results in
few men in the heavy female parties (which, since they assisted the most
in reformation are in power). This is not a bad thing as those who are
good at something should be accepted and brought into those roles. With
a political system built on accomodating feelings this makes sense, as
women _are_ generally better at doing so, while men tend to rely far
more heavily on bottom lines and rationales. The entire populace seem to have
embraced this though, as Weston cites that many spout off statistics at
a drop of a hat. To be honest, it sounds like a nation heavily
politicized and fed facts that they believe entirely and with little
skepticism, if the government says its bad for the environment then it
is boycotted. Maybe it's just me, but that seems just like the dystopia
1984, it's just a different sort of dominance.