Discipline
I saw 2 mothers today. One with a boy and one with a girl. The mother
with the daughter treated her so much better than the one with the boy.
It made me wonder about their days, how tired each of them were, what
each might be going through. what makes one mother a better one than the
other when it came to dealing with her childs requests?