The Interview narration
Worked pretty long on this one, last few days on and off. There's a
couple things going on. Which you can read one way or the other.

1. The Murderer's name, it's a reference to several notable women in
computer sciene, as well as an artificial intelligence program. This is
because, if one choses, they can interpret the woman as being a machine
of some kind, which is also supported by her manner of speech. Her
question at the end then directs the question of ethical responsibility
for whoever created her.

2. Not knowing the reference of the name, or choosing to ignore it, one
can take that the mentions to the _state_ as well as media and death
being entertainment, that the question at the end is asking who's
reponsibile for destabilization of individuals who perform terrible
acts? Because they've been desensitized and normalized to that behavior.
Which is supported by some studies

>There is now consensus that exposure to media violence is linked to
>actual violent behavior
>http://www.nytimes.com/2013/08/25/opinion/sunday/does-media-violence-lead-to-the-real-thing.html

Which all points to the story being about asking who is responsible for
the people, who without exposure to such imagery, might have led
non-violent lives.

You can also take Eliza's insistance on being sane to hold up to that
idea that people can perform terrible acts while being perfectly sane.
Even if it doesn't appear sane to others.

3. The interogator is not named. This is because he represents, to some
extent, the reader, or the state if you take case 2 as your viewpoint.
Or because he himself is an inquisition machine or an engineer who is
simply debugging or troubleshooting an issue if you take case 1. Either
way, he remains unnamed because it's not him who's important in this
story.

And those are some of the themes about this narration. Enjoy.