80s poem
Was struck by a particular graph:

	http://starlightpolitics.weebly.com/uploads/1/0/6/6/10661450/7169783.jpeg
	http://starlightpolitics.weebly.com/three-strikes-youre-out-cj-nathan--jakayla.html

From the above page showing prison population growth of a ridiculous
proportion.

	https://en.wikipedia.org/wiki/United_States_incarceration_rate#/media/File:US_incarceration_timeline-clean.svg

Reagon's war on drugs and reforms made seemed to be the biggest cause:

	http://www.justicepolicy.org/uploads/justicepolicy/documents/florida-reform-not-privatization-fact-sheet-updated.pdf

Crack/Cocaine has some interesting bits as well in the 80's:

	https://archive.is/k314m

This ^ is where 18 to 1 and the CBC (Congressional Black Caucus)
reference comes into the picture. And Cartel reference is specifically
about miami:

	https://sites.google.com/site/mjvosssunbeltpage/effects-on-miami