Flood poem
Conflicts going on for <5 years now but no one seemed to care until it
started directly affecting them and their wallets. The people crying
over how much they need to help while they're far away from the issues
that happen, and won't be the ones to deal with the direct effects. The
issues around people having to integrate and deal with conflicting
cultures and religions clashing.

crying over their wallets while demanding that the countries that deal
with the borders directly pony up for their bills while leaving them
with nothing to protect their identities with.

lots of good meaning people. but a few bad apples blow up the bunch.
people migrating not for war, but for economic reasons while taking
advantage of the systems in place to help those fleeing from war.

and the bombings and anger, and the proxied war that goes on, providing
"logistical aid" which is bloodied hands by transitive property. But
then saying they don't want to deal with the fall out.