final edit of 56 poem
overheard a girl talking about being homeless the other day and found
out there's a ticket for sleeping in public places (50$) and that the
department of economics gives her 56$ a day to live. But if she gets a
job that check goes away. So she can't work, but she cna't get out of
her situation either.

troubled life, friends leave becuase they assume she was on drugs since
she was homeless. but she says she wasn't, and that she quit drinking
but broke up with her ex because she literally fought with his mom.