Add erosion poem
Was thinking about habits and neural pathways, I remember reading
somewhere that your synapses and things get used to firing in the same
patterns and stuff. Which is partially why we sometimes have troubles
changing habits and things.

So I was thinking about each neural path as a sort of river, where the
river has created a canyon over time like erosion and it's hard to break
free of it. And how sometimes we get into it with sad thoughts as well,
and we just dig ourselves deeper and deeper into recurring feelings and
thoughts over time.