Started working on circe blog post.
Not done yet, but I should probably go to bed since I need to be off to
work  in 8 hours. Sigh. I wonder why it's so much easier to be
productive at night? Is it just that there's no one around to distract
you? I mean, I don't think it's that since the whole internet is here to
distract me, but still. Maybe it's just that when my brain is busy
trying to tell me to go to sleep it doesn't think about other things and
it makes it easier to focus?

Oh who knows.